@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.3+18
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.6 <3.0.0"
 
 dependencies:
   flutter:
@@ -29,42 +29,45 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.5
+
+  flutter_riverpod: ^2.1.1
+
+  moor_flutter: ^4.1.0
+  shared_preferences: ^2.0.15
+
   intl: ^0.17.0
-  moor_flutter: ^4.0.0
-  flutter_riverpod: ^1.0.1
-  fl_chart: ^0.40.2
-  dotted_border: ^2.0.0+1
-  percent_indicator: ^3.4.0
-  shared_preferences: ^2.0.9
-  rxdart: ^0.27.3
-  tuple: ^2.0.0
-  collection: ^1.15.0
+  rxdart: ^0.27.7
+  tuple: ^2.0.1
+  collection: ^1.16.0
+  freezed_annotation: ^2.2.0
+  numeral: ^2.0.1
   package_info: ^2.0.2
+  fimber: ^0.6.6
+
+  fl_chart: ^0.55.2
+  dotted_border: ^2.0.0+3
+  percent_indicator: ^4.2.2
+  smooth_page_indicator: ^1.0.0+2
+  syncfusion_flutter_gauges: ^20.3.58
+
   firebase_core: ^2.3.0
   firebase_crashlytics: ^3.0.6
   firebase_analytics: ^10.0.6
-  fimber: ^0.6.4
-  smooth_page_indicator: ^1.0.0+2
-  numeral: ^1.2.5
-  syncfusion_flutter_gauges: ^20.3.58
 
-  flutter_cupertino_date_picker:
-    git: https://github.com/yasseralsamman/flutter-cupertino-date-picker.git
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  moor_generator: ^4.0.0
-  build_runner:
-  flutter_launcher_icons: ^0.9.0
+  flutter_lints: ^2.0.1
+  moor_generator: ^4.6.0+1
+  freezed: ^2.2.0
+  build_runner: ^2.3.2
+  flutter_launcher_icons: ^0.11.0
 
 flutter_icons:
   android: "ic_launcher"
   ios: true
   image_path: "assets/images/expense_manage.png"
-
-#dependency_overrides:
-#  intl: ^0.16.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
