@@ -10,98 +10,86 @@ part of 'app_database.dart';
 class EntryEntityData extends DataClass implements Insertable<EntryEntityData> {
   final int id;
   final double amount;
-  final int categoryId;
+  final int? categoryId;
   final DateTime modifiedDate;
   final String description;
   EntryEntityData(
-      {@required this.id,
-      @required this.amount,
+      {required this.id,
+      required this.amount,
       this.categoryId,
-      @required this.modifiedDate,
-      @required this.description});
+      required this.modifiedDate,
+      required this.description});
   factory EntryEntityData.fromData(
       Map<String, dynamic> data, GeneratedDatabase db,
-      {String prefix}) {
+      {String? prefix}) {
     final effectivePrefix = prefix ?? '';
     return EntryEntityData(
-      id: const IntType().mapFromDatabaseResponse(data['${effectivePrefix}id']),
+      id: const IntType()
+          .mapFromDatabaseResponse(data['${effectivePrefix}id'])!,
       amount: const RealType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}amount']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}amount'])!,
       categoryId: const IntType()
           .mapFromDatabaseResponse(data['${effectivePrefix}category_id']),
       modifiedDate: const DateTimeType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}modified_date']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}modified_date'])!,
       description: const StringType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}description']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}description'])!,
     );
   }
   @override
   Map<String, Expression> toColumns(bool nullToAbsent) {
     final map = <String, Expression>{};
-    if (!nullToAbsent || id != null) {
-      map['id'] = Variable<int>(id);
-    }
-    if (!nullToAbsent || amount != null) {
-      map['amount'] = Variable<double>(amount);
-    }
+    map['id'] = Variable<int>(id);
+    map['amount'] = Variable<double>(amount);
     if (!nullToAbsent || categoryId != null) {
-      map['category_id'] = Variable<int>(categoryId);
+      map['category_id'] = Variable<int?>(categoryId);
     }
-    if (!nullToAbsent || modifiedDate != null) {
-      map['modified_date'] = Variable<DateTime>(modifiedDate);
-    }
-    if (!nullToAbsent || description != null) {
-      map['description'] = Variable<String>(description);
-    }
+    map['modified_date'] = Variable<DateTime>(modifiedDate);
+    map['description'] = Variable<String>(description);
     return map;
   }
 
   EntryEntityCompanion toCompanion(bool nullToAbsent) {
     return EntryEntityCompanion(
-      id: id == null && nullToAbsent ? const Value.absent() : Value(id),
-      amount:
-          amount == null && nullToAbsent ? const Value.absent() : Value(amount),
+      id: Value(id),
+      amount: Value(amount),
       categoryId: categoryId == null && nullToAbsent
           ? const Value.absent()
           : Value(categoryId),
-      modifiedDate: modifiedDate == null && nullToAbsent
-          ? const Value.absent()
-          : Value(modifiedDate),
-      description: description == null && nullToAbsent
-          ? const Value.absent()
-          : Value(description),
+      modifiedDate: Value(modifiedDate),
+      description: Value(description),
     );
   }
 
   factory EntryEntityData.fromJson(Map<String, dynamic> json,
-      {ValueSerializer serializer}) {
+      {ValueSerializer? serializer}) {
     serializer ??= moorRuntimeOptions.defaultSerializer;
     return EntryEntityData(
       id: serializer.fromJson<int>(json['id']),
       amount: serializer.fromJson<double>(json['amount']),
-      categoryId: serializer.fromJson<int>(json['categoryId']),
+      categoryId: serializer.fromJson<int?>(json['categoryId']),
       modifiedDate: serializer.fromJson<DateTime>(json['modifiedDate']),
       description: serializer.fromJson<String>(json['description']),
     );
   }
   @override
-  Map<String, dynamic> toJson({ValueSerializer serializer}) {
+  Map<String, dynamic> toJson({ValueSerializer? serializer}) {
     serializer ??= moorRuntimeOptions.defaultSerializer;
     return <String, dynamic>{
       'id': serializer.toJson<int>(id),
       'amount': serializer.toJson<double>(amount),
-      'categoryId': serializer.toJson<int>(categoryId),
+      'categoryId': serializer.toJson<int?>(categoryId),
       'modifiedDate': serializer.toJson<DateTime>(modifiedDate),
       'description': serializer.toJson<String>(description),
     };
   }
 
   EntryEntityData copyWith(
-          {int id,
-          double amount,
-          int categoryId,
-          DateTime modifiedDate,
-          String description}) =>
+          {int? id,
+          double? amount,
+          int? categoryId,
+          DateTime? modifiedDate,
+          String? description}) =>
       EntryEntityData(
         id: id ?? this.id,
         amount: amount ?? this.amount,
@@ -138,7 +126,7 @@ class EntryEntityData extends DataClass implements Insertable<EntryEntityData> {
 class EntryEntityCompanion extends UpdateCompanion<EntryEntityData> {
   final Value<int> id;
   final Value<double> amount;
-  final Value<int> categoryId;
+  final Value<int?> categoryId;
   final Value<DateTime> modifiedDate;
   final Value<String> description;
   const EntryEntityCompanion({
@@ -150,19 +138,19 @@ class EntryEntityCompanion extends UpdateCompanion<EntryEntityData> {
   });
   EntryEntityCompanion.insert({
     this.id = const Value.absent(),
-    @required double amount,
+    required double amount,
     this.categoryId = const Value.absent(),
-    @required DateTime modifiedDate,
-    @required String description,
+    required DateTime modifiedDate,
+    required String description,
   })  : amount = Value(amount),
         modifiedDate = Value(modifiedDate),
         description = Value(description);
   static Insertable<EntryEntityData> custom({
-    Expression<int> id,
-    Expression<double> amount,
-    Expression<int> categoryId,
-    Expression<DateTime> modifiedDate,
-    Expression<String> description,
+    Expression<int>? id,
+    Expression<double>? amount,
+    Expression<int?>? categoryId,
+    Expression<DateTime>? modifiedDate,
+    Expression<String>? description,
   }) {
     return RawValuesInsertable({
       if (id != null) 'id': id,
@@ -174,11 +162,11 @@ class EntryEntityCompanion extends UpdateCompanion<EntryEntityData> {
   }
 
   EntryEntityCompanion copyWith(
-      {Value<int> id,
-      Value<double> amount,
-      Value<int> categoryId,
-      Value<DateTime> modifiedDate,
-      Value<String> description}) {
+      {Value<int>? id,
+      Value<double>? amount,
+      Value<int?>? categoryId,
+      Value<DateTime>? modifiedDate,
+      Value<String>? description}) {
     return EntryEntityCompanion(
       id: id ?? this.id,
       amount: amount ?? this.amount,
@@ -198,7 +186,7 @@ class EntryEntityCompanion extends UpdateCompanion<EntryEntityData> {
       map['amount'] = Variable<double>(amount.value);
     }
     if (categoryId.present) {
-      map['category_id'] = Variable<int>(categoryId.value);
+      map['category_id'] = Variable<int?>(categoryId.value);
     }
     if (modifiedDate.present) {
       map['modified_date'] = Variable<DateTime>(modifiedDate.value);
@@ -226,47 +214,42 @@ class $EntryEntityTable extends EntryEntity
     with TableInfo<$EntryEntityTable, EntryEntityData> {
   @override
   final GeneratedDatabase attachedDatabase;
-  final String _alias;
+  final String? _alias;
   $EntryEntityTable(this.attachedDatabase, [this._alias]);
   final VerificationMeta _idMeta = const VerificationMeta('id');
-  GeneratedColumn<int> _id;
   @override
-  GeneratedColumn<int> get id =>
-      _id ??= GeneratedColumn<int>('id', aliasedName, false,
-          type: const IntType(),
-          requiredDuringInsert: false,
-          defaultConstraints: 'PRIMARY KEY AUTOINCREMENT');
+  late final GeneratedColumn<int?> id = GeneratedColumn<int?>(
+      'id', aliasedName, false,
+      type: const IntType(),
+      requiredDuringInsert: false,
+      defaultConstraints: 'PRIMARY KEY AUTOINCREMENT');
   final VerificationMeta _amountMeta = const VerificationMeta('amount');
-  GeneratedColumn<double> _amount;
   @override
-  GeneratedColumn<double> get amount =>
-      _amount ??= GeneratedColumn<double>('amount', aliasedName, false,
-          type: const RealType(), requiredDuringInsert: true);
+  late final GeneratedColumn<double?> amount = GeneratedColumn<double?>(
+      'amount', aliasedName, false,
+      type: const RealType(), requiredDuringInsert: true);
   final VerificationMeta _categoryIdMeta = const VerificationMeta('categoryId');
-  GeneratedColumn<int> _categoryId;
   @override
-  GeneratedColumn<int> get categoryId =>
-      _categoryId ??= GeneratedColumn<int>('category_id', aliasedName, true,
-          type: const IntType(),
-          requiredDuringInsert: false,
-          $customConstraints:
-              'NULL REFERENCES category_entity(id) ON DELETE SET NULL');
+  late final GeneratedColumn<int?> categoryId = GeneratedColumn<int?>(
+      'category_id', aliasedName, true,
+      type: const IntType(),
+      requiredDuringInsert: false,
+      $customConstraints:
+          'NULL REFERENCES category_entity(id) ON DELETE SET NULL');
   final VerificationMeta _modifiedDateMeta =
       const VerificationMeta('modifiedDate');
-  GeneratedColumn<DateTime> _modifiedDate;
   @override
-  GeneratedColumn<DateTime> get modifiedDate => _modifiedDate ??=
-      GeneratedColumn<DateTime>('modified_date', aliasedName, false,
+  late final GeneratedColumn<DateTime?> modifiedDate =
+      GeneratedColumn<DateTime?>('modified_date', aliasedName, false,
           type: const IntType(), requiredDuringInsert: true);
   final VerificationMeta _descriptionMeta =
       const VerificationMeta('description');
-  GeneratedColumn<String> _description;
   @override
-  GeneratedColumn<String> get description => _description ??=
-      GeneratedColumn<String>('description', aliasedName, false,
-          additionalChecks: GeneratedColumn.checkTextLength(maxTextLength: 100),
-          type: const StringType(),
-          requiredDuringInsert: true);
+  late final GeneratedColumn<String?> description = GeneratedColumn<String?>(
+      'description', aliasedName, false,
+      additionalChecks: GeneratedColumn.checkTextLength(maxTextLength: 100),
+      type: const StringType(),
+      requiredDuringInsert: true);
   @override
   List<GeneratedColumn> get $columns =>
       [id, amount, categoryId, modifiedDate, description];
@@ -280,11 +263,11 @@ class $EntryEntityTable extends EntryEntity
     final context = VerificationContext();
     final data = instance.toColumns(true);
     if (data.containsKey('id')) {
-      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id'], _idMeta));
+      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id']!, _idMeta));
     }
     if (data.containsKey('amount')) {
       context.handle(_amountMeta,
-          amount.isAcceptableOrUnknown(data['amount'], _amountMeta));
+          amount.isAcceptableOrUnknown(data['amount']!, _amountMeta));
     } else if (isInserting) {
       context.missing(_amountMeta);
     }
@@ -292,13 +275,13 @@ class $EntryEntityTable extends EntryEntity
       context.handle(
           _categoryIdMeta,
           categoryId.isAcceptableOrUnknown(
-              data['category_id'], _categoryIdMeta));
+              data['category_id']!, _categoryIdMeta));
     }
     if (data.containsKey('modified_date')) {
       context.handle(
           _modifiedDateMeta,
           modifiedDate.isAcceptableOrUnknown(
-              data['modified_date'], _modifiedDateMeta));
+              data['modified_date']!, _modifiedDateMeta));
     } else if (isInserting) {
       context.missing(_modifiedDateMeta);
     }
@@ -306,7 +289,7 @@ class $EntryEntityTable extends EntryEntity
       context.handle(
           _descriptionMeta,
           description.isAcceptableOrUnknown(
-              data['description'], _descriptionMeta));
+              data['description']!, _descriptionMeta));
     } else if (isInserting) {
       context.missing(_descriptionMeta);
     }
@@ -316,7 +299,7 @@ class $EntryEntityTable extends EntryEntity
   @override
   Set<GeneratedColumn> get $primaryKey => {id};
   @override
-  EntryEntityData map(Map<String, dynamic> data, {String tablePrefix}) {
+  EntryEntityData map(Map<String, dynamic> data, {String? tablePrefix}) {
     return EntryEntityData.fromData(data, attachedDatabase,
         prefix: tablePrefix != null ? '$tablePrefix.' : null);
   }
@@ -335,64 +318,51 @@ class CategoryEntityData extends DataClass
   final String icon;
   final String iconColor;
   CategoryEntityData(
-      {@required this.id,
-      @required this.position,
-      @required this.name,
-      @required this.icon,
-      @required this.iconColor});
+      {required this.id,
+      required this.position,
+      required this.name,
+      required this.icon,
+      required this.iconColor});
   factory CategoryEntityData.fromData(
       Map<String, dynamic> data, GeneratedDatabase db,
-      {String prefix}) {
+      {String? prefix}) {
     final effectivePrefix = prefix ?? '';
     return CategoryEntityData(
-      id: const IntType().mapFromDatabaseResponse(data['${effectivePrefix}id']),
+      id: const IntType()
+          .mapFromDatabaseResponse(data['${effectivePrefix}id'])!,
       position: const IntType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}position']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}position'])!,
       name: const StringType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}name']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}name'])!,
       icon: const StringType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}icon']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}icon'])!,
       iconColor: const StringType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}icon_color']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}icon_color'])!,
     );
   }
   @override
   Map<String, Expression> toColumns(bool nullToAbsent) {
     final map = <String, Expression>{};
-    if (!nullToAbsent || id != null) {
-      map['id'] = Variable<int>(id);
-    }
-    if (!nullToAbsent || position != null) {
-      map['position'] = Variable<int>(position);
-    }
-    if (!nullToAbsent || name != null) {
-      map['name'] = Variable<String>(name);
-    }
-    if (!nullToAbsent || icon != null) {
-      map['icon'] = Variable<String>(icon);
-    }
-    if (!nullToAbsent || iconColor != null) {
-      map['icon_color'] = Variable<String>(iconColor);
-    }
+    map['id'] = Variable<int>(id);
+    map['position'] = Variable<int>(position);
+    map['name'] = Variable<String>(name);
+    map['icon'] = Variable<String>(icon);
+    map['icon_color'] = Variable<String>(iconColor);
     return map;
   }
 
   CategoryEntityCompanion toCompanion(bool nullToAbsent) {
     return CategoryEntityCompanion(
-      id: id == null && nullToAbsent ? const Value.absent() : Value(id),
-      position: position == null && nullToAbsent
-          ? const Value.absent()
-          : Value(position),
-      name: name == null && nullToAbsent ? const Value.absent() : Value(name),
-      icon: icon == null && nullToAbsent ? const Value.absent() : Value(icon),
-      iconColor: iconColor == null && nullToAbsent
-          ? const Value.absent()
-          : Value(iconColor),
+      id: Value(id),
+      position: Value(position),
+      name: Value(name),
+      icon: Value(icon),
+      iconColor: Value(iconColor),
     );
   }
 
   factory CategoryEntityData.fromJson(Map<String, dynamic> json,
-      {ValueSerializer serializer}) {
+      {ValueSerializer? serializer}) {
     serializer ??= moorRuntimeOptions.defaultSerializer;
     return CategoryEntityData(
       id: serializer.fromJson<int>(json['id']),
@@ -403,7 +373,7 @@ class CategoryEntityData extends DataClass
     );
   }
   @override
-  Map<String, dynamic> toJson({ValueSerializer serializer}) {
+  Map<String, dynamic> toJson({ValueSerializer? serializer}) {
     serializer ??= moorRuntimeOptions.defaultSerializer;
     return <String, dynamic>{
       'id': serializer.toJson<int>(id),
@@ -415,7 +385,11 @@ class CategoryEntityData extends DataClass
   }
 
   CategoryEntityData copyWith(
-          {int id, int position, String name, String icon, String iconColor}) =>
+          {int? id,
+          int? position,
+          String? name,
+          String? icon,
+          String? iconColor}) =>
       CategoryEntityData(
         id: id ?? this.id,
         position: position ?? this.position,
@@ -463,20 +437,20 @@ class CategoryEntityCompanion extends UpdateCompanion<CategoryEntityData> {
   });
   CategoryEntityCompanion.insert({
     this.id = const Value.absent(),
-    @required int position,
-    @required String name,
-    @required String icon,
-    @required String iconColor,
+    required int position,
+    required String name,
+    required String icon,
+    required String iconColor,
   })  : position = Value(position),
         name = Value(name),
         icon = Value(icon),
         iconColor = Value(iconColor);
   static Insertable<CategoryEntityData> custom({
-    Expression<int> id,
-    Expression<int> position,
-    Expression<String> name,
-    Expression<String> icon,
-    Expression<String> iconColor,
+    Expression<int>? id,
+    Expression<int>? position,
+    Expression<String>? name,
+    Expression<String>? icon,
+    Expression<String>? iconColor,
   }) {
     return RawValuesInsertable({
       if (id != null) 'id': id,
@@ -488,11 +462,11 @@ class CategoryEntityCompanion extends UpdateCompanion<CategoryEntityData> {
   }
 
   CategoryEntityCompanion copyWith(
-      {Value<int> id,
-      Value<int> position,
-      Value<String> name,
-      Value<String> icon,
-      Value<String> iconColor}) {
+      {Value<int>? id,
+      Value<int>? position,
+      Value<String>? name,
+      Value<String>? icon,
+      Value<String>? iconColor}) {
     return CategoryEntityCompanion(
       id: id ?? this.id,
       position: position ?? this.position,
@@ -540,43 +514,38 @@ class $CategoryEntityTable extends CategoryEntity
     with TableInfo<$CategoryEntityTable, CategoryEntityData> {
   @override
   final GeneratedDatabase attachedDatabase;
-  final String _alias;
+  final String? _alias;
   $CategoryEntityTable(this.attachedDatabase, [this._alias]);
   final VerificationMeta _idMeta = const VerificationMeta('id');
-  GeneratedColumn<int> _id;
   @override
-  GeneratedColumn<int> get id =>
-      _id ??= GeneratedColumn<int>('id', aliasedName, false,
-          type: const IntType(),
-          requiredDuringInsert: false,
-          defaultConstraints: 'PRIMARY KEY AUTOINCREMENT');
+  late final GeneratedColumn<int?> id = GeneratedColumn<int?>(
+      'id', aliasedName, false,
+      type: const IntType(),
+      requiredDuringInsert: false,
+      defaultConstraints: 'PRIMARY KEY AUTOINCREMENT');
   final VerificationMeta _positionMeta = const VerificationMeta('position');
-  GeneratedColumn<int> _position;
   @override
-  GeneratedColumn<int> get position =>
-      _position ??= GeneratedColumn<int>('position', aliasedName, false,
-          type: const IntType(), requiredDuringInsert: true);
+  late final GeneratedColumn<int?> position = GeneratedColumn<int?>(
+      'position', aliasedName, false,
+      type: const IntType(), requiredDuringInsert: true);
   final VerificationMeta _nameMeta = const VerificationMeta('name');
-  GeneratedColumn<String> _name;
   @override
-  GeneratedColumn<String> get name => _name ??= GeneratedColumn<String>(
+  late final GeneratedColumn<String?> name = GeneratedColumn<String?>(
       'name', aliasedName, false,
       additionalChecks:
           GeneratedColumn.checkTextLength(minTextLength: 3, maxTextLength: 20),
       type: const StringType(),
       requiredDuringInsert: true);
   final VerificationMeta _iconMeta = const VerificationMeta('icon');
-  GeneratedColumn<String> _icon;
   @override
-  GeneratedColumn<String> get icon =>
-      _icon ??= GeneratedColumn<String>('icon', aliasedName, false,
-          type: const StringType(), requiredDuringInsert: true);
+  late final GeneratedColumn<String?> icon = GeneratedColumn<String?>(
+      'icon', aliasedName, false,
+      type: const StringType(), requiredDuringInsert: true);
   final VerificationMeta _iconColorMeta = const VerificationMeta('iconColor');
-  GeneratedColumn<String> _iconColor;
   @override
-  GeneratedColumn<String> get iconColor =>
-      _iconColor ??= GeneratedColumn<String>('icon_color', aliasedName, false,
-          type: const StringType(), requiredDuringInsert: true);
+  late final GeneratedColumn<String?> iconColor = GeneratedColumn<String?>(
+      'icon_color', aliasedName, false,
+      type: const StringType(), requiredDuringInsert: true);
   @override
   List<GeneratedColumn> get $columns => [id, position, name, icon, iconColor];
   @override
@@ -589,29 +558,29 @@ class $CategoryEntityTable extends CategoryEntity
     final context = VerificationContext();
     final data = instance.toColumns(true);
     if (data.containsKey('id')) {
-      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id'], _idMeta));
+      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id']!, _idMeta));
     }
     if (data.containsKey('position')) {
       context.handle(_positionMeta,
-          position.isAcceptableOrUnknown(data['position'], _positionMeta));
+          position.isAcceptableOrUnknown(data['position']!, _positionMeta));
     } else if (isInserting) {
       context.missing(_positionMeta);
     }
     if (data.containsKey('name')) {
       context.handle(
-          _nameMeta, name.isAcceptableOrUnknown(data['name'], _nameMeta));
+          _nameMeta, name.isAcceptableOrUnknown(data['name']!, _nameMeta));
     } else if (isInserting) {
       context.missing(_nameMeta);
     }
     if (data.containsKey('icon')) {
       context.handle(
-          _iconMeta, icon.isAcceptableOrUnknown(data['icon'], _iconMeta));
+          _iconMeta, icon.isAcceptableOrUnknown(data['icon']!, _iconMeta));
     } else if (isInserting) {
       context.missing(_iconMeta);
     }
     if (data.containsKey('icon_color')) {
       context.handle(_iconColorMeta,
-          iconColor.isAcceptableOrUnknown(data['icon_color'], _iconColorMeta));
+          iconColor.isAcceptableOrUnknown(data['icon_color']!, _iconColorMeta));
     } else if (isInserting) {
       context.missing(_iconColorMeta);
     }
@@ -621,7 +590,7 @@ class $CategoryEntityTable extends CategoryEntity
   @override
   Set<GeneratedColumn> get $primaryKey => {id};
   @override
-  CategoryEntityData map(Map<String, dynamic> data, {String tablePrefix}) {
+  CategoryEntityData map(Map<String, dynamic> data, {String? tablePrefix}) {
     return CategoryEntityData.fromData(data, attachedDatabase,
         prefix: tablePrefix != null ? '$tablePrefix.' : null);
   }
@@ -640,64 +609,51 @@ class IncomeCategoryEntityData extends DataClass
   final String icon;
   final String iconColor;
   IncomeCategoryEntityData(
-      {@required this.id,
-      @required this.position,
-      @required this.name,
-      @required this.icon,
-      @required this.iconColor});
+      {required this.id,
+      required this.position,
+      required this.name,
+      required this.icon,
+      required this.iconColor});
   factory IncomeCategoryEntityData.fromData(
       Map<String, dynamic> data, GeneratedDatabase db,
-      {String prefix}) {
+      {String? prefix}) {
     final effectivePrefix = prefix ?? '';
     return IncomeCategoryEntityData(
-      id: const IntType().mapFromDatabaseResponse(data['${effectivePrefix}id']),
+      id: const IntType()
+          .mapFromDatabaseResponse(data['${effectivePrefix}id'])!,
       position: const IntType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}position']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}position'])!,
       name: const StringType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}name']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}name'])!,
       icon: const StringType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}icon']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}icon'])!,
       iconColor: const StringType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}icon_color']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}icon_color'])!,
     );
   }
   @override
   Map<String, Expression> toColumns(bool nullToAbsent) {
     final map = <String, Expression>{};
-    if (!nullToAbsent || id != null) {
-      map['id'] = Variable<int>(id);
-    }
-    if (!nullToAbsent || position != null) {
-      map['position'] = Variable<int>(position);
-    }
-    if (!nullToAbsent || name != null) {
-      map['name'] = Variable<String>(name);
-    }
-    if (!nullToAbsent || icon != null) {
-      map['icon'] = Variable<String>(icon);
-    }
-    if (!nullToAbsent || iconColor != null) {
-      map['icon_color'] = Variable<String>(iconColor);
-    }
+    map['id'] = Variable<int>(id);
+    map['position'] = Variable<int>(position);
+    map['name'] = Variable<String>(name);
+    map['icon'] = Variable<String>(icon);
+    map['icon_color'] = Variable<String>(iconColor);
     return map;
   }
 
   IncomeCategoryEntityCompanion toCompanion(bool nullToAbsent) {
     return IncomeCategoryEntityCompanion(
-      id: id == null && nullToAbsent ? const Value.absent() : Value(id),
-      position: position == null && nullToAbsent
-          ? const Value.absent()
-          : Value(position),
-      name: name == null && nullToAbsent ? const Value.absent() : Value(name),
-      icon: icon == null && nullToAbsent ? const Value.absent() : Value(icon),
-      iconColor: iconColor == null && nullToAbsent
-          ? const Value.absent()
-          : Value(iconColor),
+      id: Value(id),
+      position: Value(position),
+      name: Value(name),
+      icon: Value(icon),
+      iconColor: Value(iconColor),
     );
   }
 
   factory IncomeCategoryEntityData.fromJson(Map<String, dynamic> json,
-      {ValueSerializer serializer}) {
+      {ValueSerializer? serializer}) {
     serializer ??= moorRuntimeOptions.defaultSerializer;
     return IncomeCategoryEntityData(
       id: serializer.fromJson<int>(json['id']),
@@ -708,7 +664,7 @@ class IncomeCategoryEntityData extends DataClass
     );
   }
   @override
-  Map<String, dynamic> toJson({ValueSerializer serializer}) {
+  Map<String, dynamic> toJson({ValueSerializer? serializer}) {
     serializer ??= moorRuntimeOptions.defaultSerializer;
     return <String, dynamic>{
       'id': serializer.toJson<int>(id),
@@ -720,7 +676,11 @@ class IncomeCategoryEntityData extends DataClass
   }
 
   IncomeCategoryEntityData copyWith(
-          {int id, int position, String name, String icon, String iconColor}) =>
+          {int? id,
+          int? position,
+          String? name,
+          String? icon,
+          String? iconColor}) =>
       IncomeCategoryEntityData(
         id: id ?? this.id,
         position: position ?? this.position,
@@ -769,20 +729,20 @@ class IncomeCategoryEntityCompanion
   });
   IncomeCategoryEntityCompanion.insert({
     this.id = const Value.absent(),
-    @required int position,
-    @required String name,
-    @required String icon,
-    @required String iconColor,
+    required int position,
+    required String name,
+    required String icon,
+    required String iconColor,
   })  : position = Value(position),
         name = Value(name),
         icon = Value(icon),
         iconColor = Value(iconColor);
   static Insertable<IncomeCategoryEntityData> custom({
-    Expression<int> id,
-    Expression<int> position,
-    Expression<String> name,
-    Expression<String> icon,
-    Expression<String> iconColor,
+    Expression<int>? id,
+    Expression<int>? position,
+    Expression<String>? name,
+    Expression<String>? icon,
+    Expression<String>? iconColor,
   }) {
     return RawValuesInsertable({
       if (id != null) 'id': id,
@@ -794,11 +754,11 @@ class IncomeCategoryEntityCompanion
   }
 
   IncomeCategoryEntityCompanion copyWith(
-      {Value<int> id,
-      Value<int> position,
-      Value<String> name,
-      Value<String> icon,
-      Value<String> iconColor}) {
+      {Value<int>? id,
+      Value<int>? position,
+      Value<String>? name,
+      Value<String>? icon,
+      Value<String>? iconColor}) {
     return IncomeCategoryEntityCompanion(
       id: id ?? this.id,
       position: position ?? this.position,
@@ -846,43 +806,38 @@ class $IncomeCategoryEntityTable extends IncomeCategoryEntity
     with TableInfo<$IncomeCategoryEntityTable, IncomeCategoryEntityData> {
   @override
   final GeneratedDatabase attachedDatabase;
-  final String _alias;
+  final String? _alias;
   $IncomeCategoryEntityTable(this.attachedDatabase, [this._alias]);
   final VerificationMeta _idMeta = const VerificationMeta('id');
-  GeneratedColumn<int> _id;
   @override
-  GeneratedColumn<int> get id =>
-      _id ??= GeneratedColumn<int>('id', aliasedName, false,
-          type: const IntType(),
-          requiredDuringInsert: false,
-          defaultConstraints: 'PRIMARY KEY AUTOINCREMENT');
+  late final GeneratedColumn<int?> id = GeneratedColumn<int?>(
+      'id', aliasedName, false,
+      type: const IntType(),
+      requiredDuringInsert: false,
+      defaultConstraints: 'PRIMARY KEY AUTOINCREMENT');
   final VerificationMeta _positionMeta = const VerificationMeta('position');
-  GeneratedColumn<int> _position;
   @override
-  GeneratedColumn<int> get position =>
-      _position ??= GeneratedColumn<int>('position', aliasedName, false,
-          type: const IntType(), requiredDuringInsert: true);
+  late final GeneratedColumn<int?> position = GeneratedColumn<int?>(
+      'position', aliasedName, false,
+      type: const IntType(), requiredDuringInsert: true);
   final VerificationMeta _nameMeta = const VerificationMeta('name');
-  GeneratedColumn<String> _name;
   @override
-  GeneratedColumn<String> get name => _name ??= GeneratedColumn<String>(
+  late final GeneratedColumn<String?> name = GeneratedColumn<String?>(
       'name', aliasedName, false,
       additionalChecks:
           GeneratedColumn.checkTextLength(minTextLength: 3, maxTextLength: 20),
       type: const StringType(),
       requiredDuringInsert: true);
   final VerificationMeta _iconMeta = const VerificationMeta('icon');
-  GeneratedColumn<String> _icon;
   @override
-  GeneratedColumn<String> get icon =>
-      _icon ??= GeneratedColumn<String>('icon', aliasedName, false,
-          type: const StringType(), requiredDuringInsert: true);
+  late final GeneratedColumn<String?> icon = GeneratedColumn<String?>(
+      'icon', aliasedName, false,
+      type: const StringType(), requiredDuringInsert: true);
   final VerificationMeta _iconColorMeta = const VerificationMeta('iconColor');
-  GeneratedColumn<String> _iconColor;
   @override
-  GeneratedColumn<String> get iconColor =>
-      _iconColor ??= GeneratedColumn<String>('icon_color', aliasedName, false,
-          type: const StringType(), requiredDuringInsert: true);
+  late final GeneratedColumn<String?> iconColor = GeneratedColumn<String?>(
+      'icon_color', aliasedName, false,
+      type: const StringType(), requiredDuringInsert: true);
   @override
   List<GeneratedColumn> get $columns => [id, position, name, icon, iconColor];
   @override
@@ -896,29 +851,29 @@ class $IncomeCategoryEntityTable extends IncomeCategoryEntity
     final context = VerificationContext();
     final data = instance.toColumns(true);
     if (data.containsKey('id')) {
-      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id'], _idMeta));
+      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id']!, _idMeta));
     }
     if (data.containsKey('position')) {
       context.handle(_positionMeta,
-          position.isAcceptableOrUnknown(data['position'], _positionMeta));
+          position.isAcceptableOrUnknown(data['position']!, _positionMeta));
     } else if (isInserting) {
       context.missing(_positionMeta);
     }
     if (data.containsKey('name')) {
       context.handle(
-          _nameMeta, name.isAcceptableOrUnknown(data['name'], _nameMeta));
+          _nameMeta, name.isAcceptableOrUnknown(data['name']!, _nameMeta));
     } else if (isInserting) {
       context.missing(_nameMeta);
     }
     if (data.containsKey('icon')) {
       context.handle(
-          _iconMeta, icon.isAcceptableOrUnknown(data['icon'], _iconMeta));
+          _iconMeta, icon.isAcceptableOrUnknown(data['icon']!, _iconMeta));
     } else if (isInserting) {
       context.missing(_iconMeta);
     }
     if (data.containsKey('icon_color')) {
       context.handle(_iconColorMeta,
-          iconColor.isAcceptableOrUnknown(data['icon_color'], _iconColorMeta));
+          iconColor.isAcceptableOrUnknown(data['icon_color']!, _iconColorMeta));
     } else if (isInserting) {
       context.missing(_iconColorMeta);
     }
@@ -929,7 +884,7 @@ class $IncomeCategoryEntityTable extends IncomeCategoryEntity
   Set<GeneratedColumn> get $primaryKey => {id};
   @override
   IncomeCategoryEntityData map(Map<String, dynamic> data,
-      {String tablePrefix}) {
+      {String? tablePrefix}) {
     return IncomeCategoryEntityData.fromData(data, attachedDatabase,
         prefix: tablePrefix != null ? '$tablePrefix.' : null);
   }
@@ -944,98 +899,86 @@ class IncomeEntryEntityData extends DataClass
     implements Insertable<IncomeEntryEntityData> {
   final int id;
   final double amount;
-  final int categoryId;
+  final int? categoryId;
   final DateTime modifiedDate;
   final String description;
   IncomeEntryEntityData(
-      {@required this.id,
-      @required this.amount,
+      {required this.id,
+      required this.amount,
       this.categoryId,
-      @required this.modifiedDate,
-      @required this.description});
+      required this.modifiedDate,
+      required this.description});
   factory IncomeEntryEntityData.fromData(
       Map<String, dynamic> data, GeneratedDatabase db,
-      {String prefix}) {
+      {String? prefix}) {
     final effectivePrefix = prefix ?? '';
     return IncomeEntryEntityData(
-      id: const IntType().mapFromDatabaseResponse(data['${effectivePrefix}id']),
+      id: const IntType()
+          .mapFromDatabaseResponse(data['${effectivePrefix}id'])!,
       amount: const RealType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}amount']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}amount'])!,
       categoryId: const IntType()
           .mapFromDatabaseResponse(data['${effectivePrefix}category_id']),
       modifiedDate: const DateTimeType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}modified_date']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}modified_date'])!,
       description: const StringType()
-          .mapFromDatabaseResponse(data['${effectivePrefix}description']),
+          .mapFromDatabaseResponse(data['${effectivePrefix}description'])!,
     );
   }
   @override
   Map<String, Expression> toColumns(bool nullToAbsent) {
     final map = <String, Expression>{};
-    if (!nullToAbsent || id != null) {
-      map['id'] = Variable<int>(id);
-    }
-    if (!nullToAbsent || amount != null) {
-      map['amount'] = Variable<double>(amount);
-    }
+    map['id'] = Variable<int>(id);
+    map['amount'] = Variable<double>(amount);
     if (!nullToAbsent || categoryId != null) {
-      map['category_id'] = Variable<int>(categoryId);
+      map['category_id'] = Variable<int?>(categoryId);
     }
-    if (!nullToAbsent || modifiedDate != null) {
-      map['modified_date'] = Variable<DateTime>(modifiedDate);
-    }
-    if (!nullToAbsent || description != null) {
-      map['description'] = Variable<String>(description);
-    }
+    map['modified_date'] = Variable<DateTime>(modifiedDate);
+    map['description'] = Variable<String>(description);
     return map;
   }
 
   IncomeEntryEntityCompanion toCompanion(bool nullToAbsent) {
     return IncomeEntryEntityCompanion(
-      id: id == null && nullToAbsent ? const Value.absent() : Value(id),
-      amount:
-          amount == null && nullToAbsent ? const Value.absent() : Value(amount),
+      id: Value(id),
+      amount: Value(amount),
       categoryId: categoryId == null && nullToAbsent
           ? const Value.absent()
           : Value(categoryId),
-      modifiedDate: modifiedDate == null && nullToAbsent
-          ? const Value.absent()
-          : Value(modifiedDate),
-      description: description == null && nullToAbsent
-          ? const Value.absent()
-          : Value(description),
+      modifiedDate: Value(modifiedDate),
+      description: Value(description),
     );
   }
 
   factory IncomeEntryEntityData.fromJson(Map<String, dynamic> json,
-      {ValueSerializer serializer}) {
+      {ValueSerializer? serializer}) {
     serializer ??= moorRuntimeOptions.defaultSerializer;
     return IncomeEntryEntityData(
       id: serializer.fromJson<int>(json['id']),
       amount: serializer.fromJson<double>(json['amount']),
-      categoryId: serializer.fromJson<int>(json['categoryId']),
+      categoryId: serializer.fromJson<int?>(json['categoryId']),
       modifiedDate: serializer.fromJson<DateTime>(json['modifiedDate']),
       description: serializer.fromJson<String>(json['description']),
     );
   }
   @override
-  Map<String, dynamic> toJson({ValueSerializer serializer}) {
+  Map<String, dynamic> toJson({ValueSerializer? serializer}) {
     serializer ??= moorRuntimeOptions.defaultSerializer;
     return <String, dynamic>{
       'id': serializer.toJson<int>(id),
       'amount': serializer.toJson<double>(amount),
-      'categoryId': serializer.toJson<int>(categoryId),
+      'categoryId': serializer.toJson<int?>(categoryId),
       'modifiedDate': serializer.toJson<DateTime>(modifiedDate),
       'description': serializer.toJson<String>(description),
     };
   }
 
   IncomeEntryEntityData copyWith(
-          {int id,
-          double amount,
-          int categoryId,
-          DateTime modifiedDate,
-          String description}) =>
+          {int? id,
+          double? amount,
+          int? categoryId,
+          DateTime? modifiedDate,
+          String? description}) =>
       IncomeEntryEntityData(
         id: id ?? this.id,
         amount: amount ?? this.amount,
@@ -1073,7 +1016,7 @@ class IncomeEntryEntityCompanion
     extends UpdateCompanion<IncomeEntryEntityData> {
   final Value<int> id;
   final Value<double> amount;
-  final Value<int> categoryId;
+  final Value<int?> categoryId;
   final Value<DateTime> modifiedDate;
   final Value<String> description;
   const IncomeEntryEntityCompanion({
@@ -1085,19 +1028,19 @@ class IncomeEntryEntityCompanion
   });
   IncomeEntryEntityCompanion.insert({
     this.id = const Value.absent(),
-    @required double amount,
+    required double amount,
     this.categoryId = const Value.absent(),
-    @required DateTime modifiedDate,
-    @required String description,
+    required DateTime modifiedDate,
+    required String description,
   })  : amount = Value(amount),
         modifiedDate = Value(modifiedDate),
         description = Value(description);
   static Insertable<IncomeEntryEntityData> custom({
-    Expression<int> id,
-    Expression<double> amount,
-    Expression<int> categoryId,
-    Expression<DateTime> modifiedDate,
-    Expression<String> description,
+    Expression<int>? id,
+    Expression<double>? amount,
+    Expression<int?>? categoryId,
+    Expression<DateTime>? modifiedDate,
+    Expression<String>? description,
   }) {
     return RawValuesInsertable({
       if (id != null) 'id': id,
@@ -1109,11 +1052,11 @@ class IncomeEntryEntityCompanion
   }
 
   IncomeEntryEntityCompanion copyWith(
-      {Value<int> id,
-      Value<double> amount,
-      Value<int> categoryId,
-      Value<DateTime> modifiedDate,
-      Value<String> description}) {
+      {Value<int>? id,
+      Value<double>? amount,
+      Value<int?>? categoryId,
+      Value<DateTime>? modifiedDate,
+      Value<String>? description}) {
     return IncomeEntryEntityCompanion(
       id: id ?? this.id,
       amount: amount ?? this.amount,
@@ -1133,7 +1076,7 @@ class IncomeEntryEntityCompanion
       map['amount'] = Variable<double>(amount.value);
     }
     if (categoryId.present) {
-      map['category_id'] = Variable<int>(categoryId.value);
+      map['category_id'] = Variable<int?>(categoryId.value);
     }
     if (modifiedDate.present) {
       map['modified_date'] = Variable<DateTime>(modifiedDate.value);
@@ -1161,47 +1104,42 @@ class $IncomeEntryEntityTable extends IncomeEntryEntity
     with TableInfo<$IncomeEntryEntityTable, IncomeEntryEntityData> {
   @override
   final GeneratedDatabase attachedDatabase;
-  final String _alias;
+  final String? _alias;
   $IncomeEntryEntityTable(this.attachedDatabase, [this._alias]);
   final VerificationMeta _idMeta = const VerificationMeta('id');
-  GeneratedColumn<int> _id;
   @override
-  GeneratedColumn<int> get id =>
-      _id ??= GeneratedColumn<int>('id', aliasedName, false,
-          type: const IntType(),
-          requiredDuringInsert: false,
-          defaultConstraints: 'PRIMARY KEY AUTOINCREMENT');
+  late final GeneratedColumn<int?> id = GeneratedColumn<int?>(
+      'id', aliasedName, false,
+      type: const IntType(),
+      requiredDuringInsert: false,
+      defaultConstraints: 'PRIMARY KEY AUTOINCREMENT');
   final VerificationMeta _amountMeta = const VerificationMeta('amount');
-  GeneratedColumn<double> _amount;
   @override
-  GeneratedColumn<double> get amount =>
-      _amount ??= GeneratedColumn<double>('amount', aliasedName, false,
-          type: const RealType(), requiredDuringInsert: true);
+  late final GeneratedColumn<double?> amount = GeneratedColumn<double?>(
+      'amount', aliasedName, false,
+      type: const RealType(), requiredDuringInsert: true);
   final VerificationMeta _categoryIdMeta = const VerificationMeta('categoryId');
-  GeneratedColumn<int> _categoryId;
   @override
-  GeneratedColumn<int> get categoryId =>
-      _categoryId ??= GeneratedColumn<int>('category_id', aliasedName, true,
-          type: const IntType(),
-          requiredDuringInsert: false,
-          $customConstraints:
-              'NULL REFERENCES income_category_entity(id) ON DELETE SET NULL');
+  late final GeneratedColumn<int?> categoryId = GeneratedColumn<int?>(
+      'category_id', aliasedName, true,
+      type: const IntType(),
+      requiredDuringInsert: false,
+      $customConstraints:
+          'NULL REFERENCES income_category_entity(id) ON DELETE SET NULL');
   final VerificationMeta _modifiedDateMeta =
       const VerificationMeta('modifiedDate');
-  GeneratedColumn<DateTime> _modifiedDate;
   @override
-  GeneratedColumn<DateTime> get modifiedDate => _modifiedDate ??=
-      GeneratedColumn<DateTime>('modified_date', aliasedName, false,
+  late final GeneratedColumn<DateTime?> modifiedDate =
+      GeneratedColumn<DateTime?>('modified_date', aliasedName, false,
           type: const IntType(), requiredDuringInsert: true);
   final VerificationMeta _descriptionMeta =
       const VerificationMeta('description');
-  GeneratedColumn<String> _description;
   @override
-  GeneratedColumn<String> get description => _description ??=
-      GeneratedColumn<String>('description', aliasedName, false,
-          additionalChecks: GeneratedColumn.checkTextLength(maxTextLength: 100),
-          type: const StringType(),
-          requiredDuringInsert: true);
+  late final GeneratedColumn<String?> description = GeneratedColumn<String?>(
+      'description', aliasedName, false,
+      additionalChecks: GeneratedColumn.checkTextLength(maxTextLength: 100),
+      type: const StringType(),
+      requiredDuringInsert: true);
   @override
   List<GeneratedColumn> get $columns =>
       [id, amount, categoryId, modifiedDate, description];
@@ -1216,11 +1154,11 @@ class $IncomeEntryEntityTable extends IncomeEntryEntity
     final context = VerificationContext();
     final data = instance.toColumns(true);
     if (data.containsKey('id')) {
-      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id'], _idMeta));
+      context.handle(_idMeta, id.isAcceptableOrUnknown(data['id']!, _idMeta));
     }
     if (data.containsKey('amount')) {
       context.handle(_amountMeta,
-          amount.isAcceptableOrUnknown(data['amount'], _amountMeta));
+          amount.isAcceptableOrUnknown(data['amount']!, _amountMeta));
     } else if (isInserting) {
       context.missing(_amountMeta);
     }
@@ -1228,13 +1166,13 @@ class $IncomeEntryEntityTable extends IncomeEntryEntity
       context.handle(
           _categoryIdMeta,
           categoryId.isAcceptableOrUnknown(
-              data['category_id'], _categoryIdMeta));
+              data['category_id']!, _categoryIdMeta));
     }
     if (data.containsKey('modified_date')) {
       context.handle(
           _modifiedDateMeta,
           modifiedDate.isAcceptableOrUnknown(
-              data['modified_date'], _modifiedDateMeta));
+              data['modified_date']!, _modifiedDateMeta));
     } else if (isInserting) {
       context.missing(_modifiedDateMeta);
     }
@@ -1242,7 +1180,7 @@ class $IncomeEntryEntityTable extends IncomeEntryEntity
       context.handle(
           _descriptionMeta,
           description.isAcceptableOrUnknown(
-              data['description'], _descriptionMeta));
+              data['description']!, _descriptionMeta));
     } else if (isInserting) {
       context.missing(_descriptionMeta);
     }
@@ -1252,7 +1190,7 @@ class $IncomeEntryEntityTable extends IncomeEntryEntity
   @override
   Set<GeneratedColumn> get $primaryKey => {id};
   @override
-  IncomeEntryEntityData map(Map<String, dynamic> data, {String tablePrefix}) {
+  IncomeEntryEntityData map(Map<String, dynamic> data, {String? tablePrefix}) {
     return IncomeEntryEntityData.fromData(data, attachedDatabase,
         prefix: tablePrefix != null ? '$tablePrefix.' : null);
   }
@@ -1265,17 +1203,12 @@ class $IncomeEntryEntityTable extends IncomeEntryEntity
 
 abstract class _$AppDatabase extends GeneratedDatabase {
   _$AppDatabase(QueryExecutor e) : super(SqlTypeSystem.defaultInstance, e);
-  $EntryEntityTable _entryEntity;
-  $EntryEntityTable get entryEntity => _entryEntity ??= $EntryEntityTable(this);
-  $CategoryEntityTable _categoryEntity;
-  $CategoryEntityTable get categoryEntity =>
-      _categoryEntity ??= $CategoryEntityTable(this);
-  $IncomeCategoryEntityTable _incomeCategoryEntity;
-  $IncomeCategoryEntityTable get incomeCategoryEntity =>
-      _incomeCategoryEntity ??= $IncomeCategoryEntityTable(this);
-  $IncomeEntryEntityTable _incomeEntryEntity;
-  $IncomeEntryEntityTable get incomeEntryEntity =>
-      _incomeEntryEntity ??= $IncomeEntryEntityTable(this);
+  late final $EntryEntityTable entryEntity = $EntryEntityTable(this);
+  late final $CategoryEntityTable categoryEntity = $CategoryEntityTable(this);
+  late final $IncomeCategoryEntityTable incomeCategoryEntity =
+      $IncomeCategoryEntityTable(this);
+  late final $IncomeEntryEntityTable incomeEntryEntity =
+      $IncomeEntryEntityTable(this);
   @override
   Iterable<TableInfo> get allTables => allSchemaEntities.whereType<TableInfo>();
   @override
