@@ -1,40 +1,46 @@
 import 'package:expense_manager/core/constants.dart';
 import 'package:expense_manager/data/datasource/local/moor/app_database.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:expense_manager/extension/string_extension.dart';
 import 'package:expense_manager/extension/icon_data_extension.dart';
+import 'package:expense_manager/extension/string_extension.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:moor/moor.dart';
 
 @immutable
 class Category {
-  final int id;
-  final int position;
+  final int? id;
+  final int? position;
   final String name;
   final IconData icon;
   final Color iconColor;
-  final EntryType entryType;
+  final EntryType? entryType;
 
   Category({
     this.id,
     this.position,
-    @required this.name,
-    @required this.icon,
-    @required this.iconColor,
+    required this.name,
+    required this.icon,
+    required this.iconColor,
     this.entryType,
   });
 
-  Category copyWith(
-      {int id, int position, String name, IconData icon, Color iconColor}) {
+  Category copyWith({
+    int? id,
+    int? position,
+    String? name,
+    IconData? icon,
+    Color? iconColor,
+  }) {
     return Category(
-        id: id ?? this.id,
-        position: position ?? this.position,
-        name: name ?? this.name,
-        icon: icon ?? this.icon,
-        iconColor: iconColor ?? this.iconColor);
+      id: id ?? this.id,
+      position: position ?? this.position,
+      name: name ?? this.name,
+      icon: icon ?? this.icon,
+      iconColor: iconColor ?? this.iconColor,
+    );
   }
 
   factory Category.fromExpenseCategoryEntity(
-      CategoryEntityData categoryEntityData) {
+      CategoryEntityData? categoryEntityData) {
     return Category(
         id: categoryEntityData?.id,
         position: categoryEntityData?.position,
@@ -42,12 +48,13 @@ class Category {
         icon: categoryEntityData?.icon?.jsonToIconData() ??
             AppConstants.otherCategory.icon,
         iconColor: categoryEntityData?.iconColor != null
-            ? Color(int.parse(categoryEntityData?.iconColor))
+            ? Color(int.parse(categoryEntityData!.iconColor))
             : AppConstants.otherCategory.iconColor);
   }
 
   factory Category.fromIncomeCategoryEntity(
-      IncomeCategoryEntityData incomeCategoryEntityData) {
+    IncomeCategoryEntityData? incomeCategoryEntityData,
+  ) {
     return Category(
         id: incomeCategoryEntityData?.id,
         position: incomeCategoryEntityData?.position,
@@ -55,12 +62,14 @@ class Category {
         icon: incomeCategoryEntityData?.icon?.jsonToIconData() ??
             AppConstants.otherCategory.icon,
         iconColor: incomeCategoryEntityData?.iconColor != null
-            ? Color(int.parse(incomeCategoryEntityData?.iconColor))
+            ? Color(int.parse(incomeCategoryEntityData!.iconColor))
             : AppConstants.otherCategory.iconColor);
   }
 
   factory Category.fromAllCategoryEntity(
-      CategoryEntityData categoryEntityData, int entryType) {
+    CategoryEntityData? categoryEntityData,
+    int entryType,
+  ) {
     return Category(
         id: categoryEntityData?.id,
         position: categoryEntityData?.position,
@@ -68,15 +77,15 @@ class Category {
         icon: categoryEntityData?.icon?.jsonToIconData() ??
             AppConstants.otherCategory.icon,
         iconColor: categoryEntityData?.iconColor != null
-            ? Color(int.parse(categoryEntityData?.iconColor))
+            ? Color(int.parse(categoryEntityData!.iconColor))
             : AppConstants.otherCategory.iconColor,
         entryType: EntryType.values[entryType]);
   }
 
   CategoryEntityCompanion toCategoryEntityCompanion() {
     return CategoryEntityCompanion(
-        id: id == null ? Value.absent() : Value(id),
-        position: position == null ? Value.absent() : Value(position),
+        id: id == null ? Value.absent() : Value(id!),
+        position: position == null ? Value.absent() : Value(position!),
         name: Value(name),
         icon: Value(icon.iconDataToJson()),
         iconColor:
@@ -85,8 +94,8 @@ class Category {
 
   IncomeCategoryEntityCompanion toIncomeCategoryEntityCompanion() {
     return IncomeCategoryEntityCompanion(
-        id: id == null ? Value.absent() : Value(id),
-        position: position == null ? Value.absent() : Value(position),
+        id: id == null ? Value.absent() : Value(id!),
+        position: position == null ? Value.absent() : Value(position!),
         name: Value(name),
         icon: Value(icon.iconDataToJson()),
         iconColor:
@@ -95,8 +104,8 @@ class Category {
 
   CategoryEntityData toCategoryEntityData() {
     return CategoryEntityData(
-        id: id,
-        position: position,
+        id: id!,
+        position: position!,
         name: name,
         icon: icon.iconDataToJson(),
         iconColor: "0x${iconColor.value.toRadixString(16).padLeft(8, '0')}");
