@@ -1,6 +1,5 @@
 import 'package:expense_manager/data/models/category.dart';
 import 'package:expense_manager/data/models/entry.dart';
-import 'package:flutter/material.dart';
 
 class CategoryWithEntryList {
   final Category category;
@@ -8,13 +7,16 @@ class CategoryWithEntryList {
   final List<Entry> entry;
 
   CategoryWithEntryList({
-    @required this.category,
-    @required this.total,
-    @required this.entry,
+    required this.category,
+    required this.total,
+    required this.entry,
   });
 
-  CategoryWithEntryList copyWith(
-      {List<Entry> entry, double total, Category category}) {
+  CategoryWithEntryList copyWith({
+    List<Entry>? entry,
+    double? total,
+    Category? category,
+  }) {
     return CategoryWithEntryList(
       category: category ?? this.category,
       total: total ?? this.total,
