@@ -1,16 +1,18 @@
 import 'package:expense_manager/data/models/entry_with_category.dart';
-import 'package:flutter/material.dart';
 
 class History {
   final String title;
   final List<EntryWithCategory> list;
 
   History({
-    @required this.title,
-    @required this.list,
+    required this.title,
+    required this.list,
   });
 
-  History copyWith({String title, List<EntryWithCategory> list}) {
+  History copyWith({
+    String? title,
+    List<EntryWithCategory>? list,
+  }) {
     return History(title: title ?? this.title, list: list ?? this.list);
   }
 
