@@ -10,13 +10,14 @@ import 'package:tuple/tuple.dart';
 class AddCategory extends ConsumerWidget {
   final Tuple2<EntryType, Category> tuple2;
 
-  AddCategory({@required this.tuple2}) : super();
+  AddCategory({required this.tuple2}) : super();
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final vm = ref.watch(addCategoryModelProvider(tuple2));
     return Scaffold(
       appBar: AppBar(
+
         leading: InkWell(
           onTap: () {
             Navigator.pop(context);
@@ -24,18 +25,20 @@ class AddCategory extends ConsumerWidget {
           child: Icon(Icons.arrow_back_ios_rounded),
         ),
         title: DottedBorder(
-          color: Theme.of(context).appBarTheme.titleTextStyle.color,
+          color: Theme.of(context).appBarTheme.titleTextStyle!.color!,
           dashPattern: [5, 5],
           radius: Radius.circular(12),
           borderType: BorderType.RRect,
           child: Padding(
             padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
             child: Text(
-                AppLocalization.of(context).getTranslatedVal("new_category"),
-              style: Theme.of(context).appBarTheme.titleTextStyle,),
+              AppLocalization.of(context).getTranslatedVal("new_category"),
+              style: Theme.of(context).appBarTheme.titleTextStyle,
+            ),
           ),
         ),
         actions: [
+
           InkWell(
             onTap: () {
               FocusScope.of(context).unfocus();
@@ -72,7 +75,7 @@ class AddCategory extends ConsumerWidget {
               padding: const EdgeInsets.all(20),
               child: Text(
                 AppLocalization.of(context).getTranslatedVal("save"),
-                style: Theme.of(context).textTheme.subtitle2.copyWith(
+                style: Theme.of(context).textTheme.subtitle2!.copyWith(
                     fontWeight: FontWeight.bold, color: Color(0xff2196F3)),
               ),
             ),
@@ -142,7 +145,7 @@ class AddCategory extends ConsumerWidget {
                 AppLocalization.of(context).getTranslatedVal("icon"),
                 style: Theme.of(context)
                     .textTheme
-                    .subtitle2
+                    .subtitle2!
                     .copyWith(fontSize: 18, fontWeight: FontWeight.bold),
               ),
             ),
@@ -175,7 +178,7 @@ class AddCategory extends ConsumerWidget {
                 AppLocalization.of(context).getTranslatedVal("color"),
                 style: Theme.of(context)
                     .textTheme
-                    .subtitle2
+                    .subtitle2!
                     .copyWith(fontSize: 18, fontWeight: FontWeight.bold),
               ),
             ),
