@@ -5,6 +5,26 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:tuple/tuple.dart';
 
+//
+// part 'addCategory_state.freezed.dart';
+//
+// @freezed
+// class AddCategoryEntity with _$AddCategoryEntity {
+//   factory AddCategoryEntity({
+//     @required cat.Category category,
+//     @required String name,
+//     @required IconData iconData,
+//     @required Color color,
+//     @required EntryType entryType,
+//   }) = _AddCategoryEntity;
+//
+//   factory AddCategoryEntity.initial({@required cat.Category category,}) =>
+//       AddCategoryEntity(
+//           category:category,
+//         name:
+//       );
+// }
+
 final addCategoryModelProvider = ChangeNotifierProvider.autoDispose
     .family<AddCategoryViewModel, Tuple2<EntryType, cat.Category>>(
         (ref, tuple2) => AddCategoryViewModel(
@@ -14,26 +34,26 @@ final addCategoryModelProvider = ChangeNotifierProvider.autoDispose
             ));
 
 class AddCategoryViewModel with ChangeNotifier {
-  EntryRepositoryImp entryDataSourceImp;
-  cat.Category category;
-  String name;
-  IconData iconData;
-  Color color;
-  EntryType entryType;
+  final EntryRepositoryImp entryDataSourceImp;
+  cat.Category? category;
+  final EntryType entryType;
+  late String name;
+  late IconData iconData;
+  late Color color;
 
   AddCategoryViewModel({
-    @required this.entryDataSourceImp,
-    @required this.category,
-    @required this.entryType,
+    required this.entryDataSourceImp,
+    this.category,
+    required this.entryType,
   }) {
     if (category == null) {
       name = "";
       iconData = AppConstants.otherCategory.icon;
       color = AppConstants.otherCategory.iconColor;
     } else {
-      name = category.name;
-      iconData = category.icon;
-      color = category.iconColor;
+      name = category!.name;
+      iconData = category!.icon;
+      color = category!.iconColor;
     }
     notifyListeners();
   }
@@ -63,18 +83,19 @@ class AddCategoryViewModel with ChangeNotifier {
           .updateCategory(
               entryType,
               cat.Category(
-                  id: category.id,
-                  position: category.position,
-                  name: name,
-                  icon: iconData,
-                  iconColor: color))
+                id: category!.id,
+                position: category!.position,
+                name: name,
+                icon: iconData,
+                iconColor: color,
+              ))
           .listen((event) {});
     }
   }
 
   void delete() {
     entryDataSourceImp
-        .deleteCategory(entryType, category.id)
+        .deleteCategory(entryType, category!.id!)
         .listen((event) {});
   }
 
