@@ -4,8 +4,8 @@ class Logger extends ProviderObserver {
   @override
   void didUpdateProvider(
     ProviderBase provider,
-    Object previousValue,
-    Object newValue,
+    Object? previousValue,
+    Object? newValue,
     ProviderContainer container,
   ) {
     print('''
