@@ -1,15 +1,16 @@
 import 'package:intl/intl.dart';
 
 extension DateTimeExtension on DateTime {
-  DateTime copyWith(
-      {int year,
-      int month,
-      int day,
-      int hour,
-      int minute,
-      int second,
-      int millisecond,
-      int microsecond}) {
+  DateTime copyWith({
+    int? year,
+    int? month,
+    int? day,
+    int? hour,
+    int? minute,
+    int? second,
+    int? millisecond,
+    int? microsecond,
+  }) {
     return DateTime(
       year ?? this.year,
       month ?? this.month,
