@@ -3,15 +3,15 @@ import 'package:fimber/fimber.dart';
 
 extension ListExtension<T> on List<EntryWithCategory> {
   Map<int, List<EntryWithCategory>> groupListsByMonth(int currentMonth) {
-    var result = <int,List<EntryWithCategory>>{};
+    var result = <int, List<EntryWithCategory>>{};
 
     for (int i = 1; i <= currentMonth; i++) {
       result[i] = <EntryWithCategory>[];
     }
-Fimber.e(result.toString());
+    Fimber.e(result.toString());
     for (var element in this) {
-      if(element.entry.modifiedDate.month<=currentMonth)
-      (result[element.entry.modifiedDate.month])..add(element);
+      if (element.entry.modifiedDate.month <= currentMonth)
+        (result[element.entry.modifiedDate.month])!..add(element);
     }
     return result;
   }
