@@ -25,7 +25,7 @@ class Setting extends ConsumerWidget {
             },
             child: Icon(Icons.arrow_back_ios_rounded)),
         title: DottedBorder(
-          color: Theme.of(context).appBarTheme.titleTextStyle.color,
+          color: Theme.of(context).appBarTheme.titleTextStyle!.color!,
           dashPattern: [5, 5],
           radius: Radius.circular(12),
           borderType: BorderType.RRect,
@@ -67,7 +67,7 @@ class Setting extends ConsumerWidget {
                     AppLocalization.of(context).getTranslatedVal("appearance"),
                     style: Theme.of(context)
                         .textTheme
-                        .subtitle2
+                        .subtitle2!
                         .copyWith(fontSize: 16),
                   ),
                   SizedBox(height: 4),
@@ -82,7 +82,7 @@ class Setting extends ConsumerWidget {
                                 .getTranslatedVal("light_theme"))),
                     style: Theme.of(context)
                         .textTheme
-                        .subtitle1
+                        .subtitle1!
                         .copyWith(fontSize: 12),
                   ),
                 ],
@@ -116,7 +116,7 @@ class Setting extends ConsumerWidget {
                         .getTranslatedVal("month_cycle_date"),
                     style: Theme.of(context)
                         .textTheme
-                        .subtitle2
+                        .subtitle2!
                         .copyWith(fontSize: 16),
                   ),
                   SizedBox(height: 4),
@@ -124,7 +124,7 @@ class Setting extends ConsumerWidget {
                     monthStartDate.date,
                     style: Theme.of(context)
                         .textTheme
-                        .subtitle1
+                        .subtitle1!
                         .copyWith(fontSize: 12),
                   ),
                 ],
@@ -157,7 +157,7 @@ class Setting extends ConsumerWidget {
                     AppLocalization.of(context).getTranslatedVal("language"),
                     style: Theme.of(context)
                         .textTheme
-                        .subtitle2
+                        .subtitle2!
                         .copyWith(fontSize: 16),
                   ),
                   SizedBox(height: 4),
@@ -168,7 +168,7 @@ class Setting extends ConsumerWidget {
                         .name,
                     style: Theme.of(context)
                         .textTheme
-                        .subtitle1
+                        .subtitle1!
                         .copyWith(fontSize: 12),
                   ),
                 ],
@@ -201,7 +201,7 @@ class Setting extends ConsumerWidget {
                     AppLocalization.of(context).getTranslatedVal("currency"),
                     style: Theme.of(context)
                         .textTheme
-                        .subtitle2
+                        .subtitle2!
                         .copyWith(fontSize: 16),
                   ),
                   SizedBox(height: 4),
@@ -209,7 +209,7 @@ class Setting extends ConsumerWidget {
                     "${NumberFormat.simpleCurrency(locale: appState.currency.item1).currencySymbol} ${appState.currency.item2}",
                     style: Theme.of(context)
                         .textTheme
-                        .subtitle1
+                        .subtitle1!
                         .copyWith(fontSize: 12),
                   ),
                 ],
@@ -225,7 +225,7 @@ class Setting extends ConsumerWidget {
                       .getTranslatedVal("expense_manager_by_nividata"),
                   style: Theme.of(context)
                       .textTheme
-                      .subtitle1
+                      .subtitle1!
                       .copyWith(fontSize: 14),
                 ),
               ),
@@ -234,7 +234,7 @@ class Setting extends ConsumerWidget {
                   "${AppLocalization.of(context).getTranslatedVal("app_version")}${appState.appVersion}",
                   style: Theme.of(context)
                       .textTheme
-                      .subtitle1
+                      .subtitle1!
                       .copyWith(fontSize: 14),
                 ),
               ),
